Improve errors
All errors passed to callbacks or thrown by this module now have a
`.sourceMapData` property that contain as much as possible of the intended
result of the function up until the error occurred.

Also, when sources are resolved, instead of failing if one of the sources fails
to be fetched, store the error in the `.sourcesContent` array and let all
requests finish. Backwards-incompatible change.